refactor(enhancement): split transform capabilities
