+++ tests/integration/enhancements-kernel.test.ts
@@ -3,29 +3,20 @@ import { describe, it } from "node:test";
 import {
   ADDRESSES,
   ALL_FEATURES,
-  COMPANION_CURSOR_ABI,
   COMPANION_CURSOR_BYTES,
   COMPANION_PARTY_BYTES,
   COMPANION_TOOLBOX_BYTES,
   CONFIG_BYTES,
   createKernel,
   CURSOR,
-  CURSOR_EDGE,
-  CURSOR_HIDDEN,
-  CURSOR_MAGIC,
-  CURSOR_UNSUPPORTED,
-  CURSOR_VALID,
   decoded,
   DETAIL,
-  expectedRgba,
   FEATURE_NATIVE_CURSOR,
   FEATURE_TARGET_READOUT,
   FEATURE_TOOLBOX_FOUNDATION,
-  fnv1a,
   installCursorGraph,
   installGameGraph,
   installPartyDetailGraph,
-  invalidCursor,
   MESSAGE_CONFIG_START,
   paintCursor,
   PARTY_DIRTY_MESSAGES,
@@ -33,12 +24,10 @@ import {
   readChangedCompanionParty,
   readChangedCompanionToolbox,
   readCompanionSnapshot,
-  readyCursor,
   readyParty,
   readyToolbox,
   rejected,
   sameCompanionToolboxState,
-  TEXTURE_KEY,
 } from "../fixtures/enhancements.ts";
 
 describe("Companion kernel", () => {
@@ -873,194 +862,4 @@ describe("Companion kernel", () => {
     assert.equal(kernel.field(CURSOR.magic), 0);
   });
 
-  it("rejects a cursor region of the wrong size, alignment, or extent", async () => {
-    const kernel = await createKernel();
-    assert.equal(kernel.init({ cursorSize: COMPANION_CURSOR_BYTES - 1 }), 0);
-    assert.equal(kernel.init({ cursorSize: COMPANION_CURSOR_BYTES + 1 }), 0);
-    assert.equal(kernel.init({ cursorSize: 64 }), 0);
-    assert.equal(kernel.init({ cursorPointer: ADDRESSES.cursor + 1 }), 0);
-    assert.equal(kernel.init({ cursorPointer: ADDRESSES.cursor + 2 }), 0);
-    // 16 MiB of memory: this region is aligned but runs past the end.
-    assert.equal(kernel.init({ cursorPointer: 0xff_f000 }), 0);
-    assert.equal(kernel.init({ cursorPointer: 0xffff_f000 }), 0);
-    // A rejected init must leave the kernel dormant.
-    kernel.tick();
-    assert.equal(kernel.field(CURSOR.magic), 0);
-    assert.equal(kernel.init(), 1);
-  });
-
-  it("publishes a validated 32x32 cursor once per distinct cursor", async () => {
-    const kernel = await createKernel();
-    const { view } = kernel;
-    installCursorGraph(view, { hotspotX: 5, hotspotY: 7 });
-    const first = paintCursor(view, 1);
-    assert.equal(kernel.init(), 1);
-
-    // The region comes from the game's allocator, so init clears it.
-    const cleared = invalidCursor(kernel.header());
-    assert.equal(cleared.status, "invalid");
-    assert.equal(cleared.reason, "cursor");
-    assert.equal(kernel.field(CURSOR.generation), 0);
-    assert.deepEqual(
-      [...kernel.payload().slice(0, 8)],
-      [0, 0, 0, 0, 0, 0, 0, 0],
-    );
-
-    kernel.tick();
-    const ready = publishedPixels(kernel.published());
-    assert.equal(ready.status, "ready");
-    assert.equal(ready.generation, 1);
-    assert.equal(ready.flags, CURSOR_VALID);
-    assert.equal(ready.hidden, false);
-    assert.equal(ready.hotspotX, 5);
-    assert.equal(ready.hotspotY, 7);
-    assert.equal(ready.pixelHash, fnv1a(first));
-    assert.equal(kernel.field(CURSOR.magic), CURSOR_MAGIC);
-    assert.equal(
-      view.getUint16(ADDRESSES.cursor + CURSOR.abi, true),
-      COMPANION_CURSOR_ABI,
-    );
-    assert.equal(
-      view.getUint16(ADDRESSES.cursor + CURSOR.byteLength, true),
-      COMPANION_CURSOR_BYTES,
-    );
-    assert.equal(kernel.field(CURSOR.width), CURSOR_EDGE);
-    assert.equal(kernel.field(CURSOR.height), CURSOR_EDGE);
-    // BGRA 0xff112233 -> R 0x11, G 0x22, B 0x33, A 0xff.
-    assert.deepEqual([...ready.pixels.slice(0, 4)], [0x11, 0x22, 0x33, 0xff]);
-    assert.deepEqual(ready.pixels, expectedRgba(first));
-
-    const sequence = kernel.field(CURSOR.sequence);
-    assert.equal(sequence % 2, 0);
-    for (let index = 0; index < 12; index += 1) kernel.tick();
-    assert.equal(kernel.field(CURSOR.generation), 1);
-    assert.equal(kernel.field(CURSOR.sequence), sequence);
-
-    const second = paintCursor(view, 2);
-    kernel.cursorEvent();
-    kernel.tick();
-    assert.equal(kernel.field(CURSOR.generation), 2);
-    assert.equal(kernel.field(CURSOR.pixelHash), fnv1a(second));
-    assert.deepEqual(publishedPixels(kernel.published()).pixels, expectedRgba(second));
-    kernel.tick();
-    kernel.tick();
-    assert.equal(kernel.field(CURSOR.generation), 2);
-
-    // Identical pixels, moved hotspot: the pixel hash cannot see this, so the
-    // published identity must carry the hotspot too.
-    view.setUint32(ADDRESSES.art + 0x04, 9, true);
-    kernel.cursorEvent();
-    kernel.tick();
-    assert.equal(kernel.field(CURSOR.generation), 3);
-    assert.equal(kernel.field(CURSOR.pixelHash), fnv1a(second));
-    assert.equal(readyCursor(kernel.header()).hotspotY, 9);
-
-    // Show/hide moves the flags only: the bitmap is unchanged, so generation
-    // holds and the renderer's CSS cache stays warm.
-    view.setInt32(ADDRESSES.showCount, -1, true);
-    kernel.tick();
-    const gone = readyCursor(kernel.header());
-    assert.equal(gone.status, "ready");
-    assert.equal(gone.flags, CURSOR_VALID | CURSOR_HIDDEN);
-    assert.equal(gone.hidden, true);
-    assert.equal(gone.generation, 3);
-    assert.deepEqual(kernel.payload(), expectedRgba(second));
-    view.setInt32(ADDRESSES.showCount, 0, true);
-    kernel.tick();
-    assert.equal(readyCursor(kernel.header()).flags, CURSOR_VALID);
-    assert.equal(kernel.field(CURSOR.generation), 3);
-    assert.deepEqual(kernel.payload(), expectedRgba(second));
-
-  });
-
-  it("never publishes an uncommitted colour buffer as a cursor", async () => {
-    const kernel = await createKernel();
-    installCursorGraph(kernel.view);
-    assert.equal(kernel.init(), 1);
-    const sequence = kernel.field(CURSOR.sequence);
-    for (let index = 0; index < 5; index += 1) kernel.tick();
-    const header = invalidCursor(kernel.header());
-    assert.equal(header.status, "invalid");
-    assert.equal(header.flags, 0);
-    assert.equal(kernel.field(CURSOR.generation), 0);
-    assert.equal(kernel.field(CURSOR.sequence), sequence);
-    assert.equal(kernel.published(), null);
-
-    paintCursor(kernel.view, 4);
-    kernel.cursorEvent();
-    kernel.tick();
-    assert.equal(kernel.header().status, "ready");
-    assert.equal(kernel.field(CURSOR.generation), 1);
-  });
-
-  it("keeps the last good pixels while the software cursor is live", async () => {
-    const kernel = await createKernel();
-    const { view } = kernel;
-    installCursorGraph(view);
-    const good = paintCursor(view, 5);
-    assert.equal(kernel.init(), 1);
-    kernel.tick();
-    assert.equal(readyCursor(kernel.header()).generation, 1);
-
-    view.setUint32(ADDRESSES.softwareModel, 1, true);
-    const replacement = paintCursor(view, 6);
-    kernel.cursorEvent();
-    for (let index = 0; index < 3; index += 1) kernel.tick();
-    const unsupported = invalidCursor(kernel.header());
-    assert.equal(unsupported.status, "invalid");
-    assert.equal(unsupported.reason, "unsupported");
-    assert.equal(unsupported.flags, CURSOR_UNSUPPORTED);
-    assert.equal(kernel.field(CURSOR.generation), 1);
-    assert.deepEqual(kernel.payload(), expectedRgba(good));
-
-    view.setUint32(ADDRESSES.softwareModel, 0, true);
-    kernel.cursorEvent();
-    kernel.tick();
-    const recovered = publishedPixels(kernel.published());
-    assert.equal(recovered.status, "ready");
-    assert.equal(recovered.generation, 2);
-    assert.deepEqual(recovered.pixels, expectedRgba(replacement));
-  });
-
-  it("clears validity for every rejected art, handle, or texture", async () => {
-    const kernel = await createKernel();
-    const { view } = kernel;
-    installCursorGraph(view);
-    const words = paintCursor(view, 7);
-    assert.equal(kernel.init(), 1);
-    kernel.tick();
-    assert.equal(readyCursor(kernel.header()).generation, 1);
-
-    const rejections: [name: string, breakGraph: () => void][] = [
-      ["texture type", () => view.setUint32(ADDRESSES.texture + 0x0c, 9, true)],
-      ["texture width", () => view.setUint32(ADDRESSES.texture + 0x14, 64, true)],
-      ["texture height", () => view.setUint32(ADDRESSES.texture + 0x18, 16, true)],
-      ["access key", () => view.setUint32(ADDRESSES.handle + 0x08, TEXTURE_KEY + 1, true)],
-      ["null art", () => view.setUint32(ADDRESSES.activeArt, 0, true)],
-      ["misaligned art", () => view.setUint32(ADDRESSES.activeArt, ADDRESSES.art + 1, true)],
-      ["hotspot x", () => view.setUint32(ADDRESSES.art + 0x00, CURSOR_EDGE, true)],
-      ["hotspot y", () => view.setUint32(ADDRESSES.art + 0x04, 0xffff_ffff, true)],
-    ];
-    let generation = 1;
-    for (const [name, breakGraph] of rejections) {
-      breakGraph();
-      kernel.cursorEvent();
-      kernel.tick();
-      const broken = invalidCursor(kernel.header());
-      assert.equal(broken.status, "invalid", name);
-      assert.equal(broken.reason, "cursor", name);
-      assert.equal(broken.flags, 0, name);
-      assert.equal(kernel.published(), null, name);
-      // Header-only: the renderer keeps rendering the last good bitmap.
-      assert.equal(kernel.field(CURSOR.generation), generation, name);
-      assert.deepEqual(kernel.payload(), expectedRgba(words), name);
-
-      installCursorGraph(view);
-      kernel.cursorEvent();
-      kernel.tick();
-      generation += 1;
-      assert.equal(kernel.header().status, "ready", name);
-      assert.equal(kernel.field(CURSOR.generation), generation, name);
-    }
-  });
 });
