+++ tests/unit/enhancement-transform.test.ts
@@ -1,5 +1,4 @@
 import assert from "node:assert/strict";
-import { createHash } from "node:crypto";
 import { describe, it } from "node:test";
 import {
   ENHANCEMENT_TRANSFORM_ABI,
@@ -10,8 +9,6 @@ import {
 import {
   enhancementConfigWords,
   ENHANCEMENT_LAYOUT_FIELDS,
-  type EnhancementOutputHashes,
-  type KnownEnhancementBuild,
 } from "../../src/main/certification/enhancement-builds.js";
 import {
   inspectEnhancementCandidate,
@@ -21,360 +18,18 @@ import {
 } from "../../src/main/certification/enhancement-transform.js";
 import { decodeEnhancementManifest } from "../../src/renderer/enhancement-manifest.js";
 import {
-  parseCode,
-  parseExports,
-  sectionById,
-  sleb,
-  splitSections,
-} from "../../src/main/core/wasm-binary.js";
-const UNSUPPORTED_ALL_CAPABILITIES: EnhancementCapabilities = Object.freeze({
-  nativeCursor: true,
-  targetObservation: true,
-  toolbox: true,
-  commands: false,
-});
-const CURSOR_ONLY: EnhancementCapabilities = Object.freeze({
-  nativeCursor: true,
-  targetObservation: false,
-  toolbox: false,
-  commands: false,
-});
-const CURSOR_TARGET: EnhancementCapabilities = Object.freeze({
-  nativeCursor: true,
-  targetObservation: true,
-  toolbox: false,
-  commands: false,
-});
-const TARGET_ONLY: EnhancementCapabilities = Object.freeze({
-  nativeCursor: false,
-  targetObservation: true,
-  toolbox: false,
-  commands: false,
-});
-const CURSOR_TOOLBOX: EnhancementCapabilities = Object.freeze({
-  nativeCursor: true,
-  targetObservation: false,
-  toolbox: true,
-  commands: false,
-});
-const NO_CAPABILITIES: EnhancementCapabilities = Object.freeze({
-  nativeCursor: false,
-  targetObservation: false,
-  toolbox: false,
-  commands: false,
-});
-const CURSOR_TOOLBOX_COMMANDS: EnhancementCapabilities = Object.freeze({
-  nativeCursor: true,
-  targetObservation: false,
-  toolbox: true,
-  commands: true,
-});
-const CURSOR_TARGET_TOOLBOX_COMMANDS: EnhancementCapabilities = Object.freeze({
-  nativeCursor: true,
-  targetObservation: true,
-  toolbox: true,
-  commands: true,
-});
-const PARTY_DIRTY_MESSAGES = Object.freeze([
-  0x1000_0038,
-  0x1000_0039,
-  0x1000_008c,
-  0x1000_0098,
-  0x1000_00c2,
-  0x1000_0111,
-  0x1000_011e,
-  0x1000_011f,
-  0x1000_0124,
-  0x1000_0126,
-] as const);
-
-const PLACEHOLDER_OUTPUTS: EnhancementOutputHashes = Object.freeze({
-  cursor: "0".repeat(64),
-  target: "0".repeat(64),
-  cursorTarget: "0".repeat(64),
-  cursorToolbox: "0".repeat(64),
-  cursorToolboxCommands: "0".repeat(64),
-  cursorTargetToolboxCommands: "0".repeat(64),
-});
-
-function uleb(value: number): number[] {
-  const out: number[] = [];
-  do {
-    let byte = value & 0x7f;
-    value >>>= 7;
-    if (value) byte |= 0x80;
-    out.push(byte);
-  } while (value);
-  return out;
-}
-
-function section(id: number, body: number[]): number[] {
-  return [id, ...uleb(body.length), ...body];
-}
-
-function moduleWithManifest(value: unknown): WebAssembly.Module {
-  const name = [...new TextEncoder().encode(ENHANCEMENT_MANIFEST_SECTION)];
-  const payload = [...new TextEncoder().encode(JSON.stringify(value))];
-  return new WebAssembly.Module(new Uint8Array([
-    ...fixture(),
-    ...section(0, [...uleb(name.length), ...name, ...payload]),
-  ]));
-}
-
-// `hookParamType` is the WebAssembly value type of the main loop's single
-// parameter: 0x7f is i32, the signature every certified build declares. A
-// caller passes another one — 0x7e is i64 — to build the module a manifest
-// does not certify, which is the only side the mismatch can come from: a
-// KnownEnhancementBuild's hookParams is the literal ["i32"] and cannot say
-// otherwise.
-function fixture(hookParamType = 0x7f): Uint8Array {
-  const type = section(1, [
-    4,
-    0x60, 1, hookParamType, 0,
-    0x60, 5, 0x7f, 0x7f, 0x7f, 0x7f, 0x7f, 0,
-    0x60, 3, 0x7f, 0x7f, 0x7f, 0,
-    0x60, 2, 0x7f, 0x7f, 0,
-  ]);
-  const env = [3, 101, 110, 118];
-  const imports = section(2, [
-    3,
-    ...env, 1, 116, 0, 0,
-    ...env, 1, 99, 0, 1,
-    ...env, 1, 117, 0, 2,
-  ]);
-  // Eight defined functions: three hooks, three commands, the distinct
-  // recurring game-thread callback, and the shared packet sender.
-  const functions = section(3, [8, 0, 1, 2, 0, 2, 3, 3, 2]);
-  const table = section(4, [1, 0x70, 1, 5, 5]);
-  const memory = section(5, [1, 1, 1, 1]);
-  const globals = section(6, [0]);
-  const tableName = [...uleb(3), 116, 98, 108];
-  const loopName = [...new TextEncoder().encode("EmscriptenExeThreadMainLoop")];
-  const cursorName = [...new TextEncoder().encode("cursor")];
-  const uiName = [...new TextEncoder().encode("ui")];
-  const frameName = [...new TextEncoder().encode("frame")];
-  const professionName = [...new TextEncoder().encode("profession")];
-  const skillName = [...new TextEncoder().encode("skill")];
-  const senderName = [...new TextEncoder().encode("sender")];
-  const memoryName = [...new TextEncoder().encode("memory")];
-  const exports = section(7, [
-    9,
-    ...tableName, 1, 0,
-    ...uleb(memoryName.length), ...memoryName, 2, 0,
-    ...uleb(loopName.length), ...loopName, 0, 3,
-    ...uleb(cursorName.length), ...cursorName, 0, 4,
-    ...uleb(uiName.length), ...uiName, 0, 5,
-    ...uleb(frameName.length), ...frameName, 0, 8,
-    ...uleb(professionName.length), ...professionName, 0, 9,
-    ...uleb(skillName.length), ...skillName, 0, 7,
-    ...uleb(senderName.length), ...senderName, 0, 10,
-  ]);
-  const mappedSegment = [0, 0x41, 1, 0x0b, 3, 4, 3, 5];
-  const frameSegment = [0, 0x41, 4, 0x0b, 1, 8];
-  const elements = section(9, [2, ...mappedSegment, ...frameSegment]);
-  const tick = [0, 0x20, 0, 0x10, 0, 0x0b];
-  const cursor = [
-    0, 0x20, 0, 0x20, 1, 0x20, 2, 0x20, 3, 0x20, 4, 0x10, 1, 0x0b,
-  ];
-  const ui = [0, 0x20, 0, 0x20, 1, 0x20, 2, 0x10, 2, 0x0b];
-  const command = [0, 0x41, 0, 0x1a, 0x20, 0, 0x10, 0, 0x0b];
-  // A compact real-shape skill-bar builder: packet words are written at 128,
-  // then handed to the shared sender unchanged.
-  const pair = [
-    0,
-    0x41, ...sleb(128), 0x41, ...sleb(93), 0x36, 2, 0,
-    0x41, ...sleb(128), 0x20, 0, 0x36, 2, 4,
-    0x41, ...sleb(128), 0x20, 1, 0x36, 2, 8,
-    ...Array.from({ length: 8 }, (_, index) => [
-      0x41, ...sleb(128),
-      0x20, 2, 0x28, 2, ...uleb(index * 4),
-      0x36, 2, ...uleb(12 + index * 4),
-    ]).flat(),
-    0x41, ...sleb(999), 0x41, ...sleb(44), 0x41, ...sleb(128), 0x10, 10,
-    0x0b,
-  ];
-  const frame = [0, 0x20, 0, 0x10, 0, 0x0b];
-  const profession = [
-    0,
-    0x41, 0, 0x41, ...sleb(65), 0x36, 2, 0,
-    0x41, 0, 0x20, 0, 0x36, 2, 4,
-    0x41, 0, 0x20, 1, 0x36, 2, 8,
-    0x41, 0xe7, 0x07, 0x41, 12, 0x41, 0, 0x10, 10,
-    0x0b,
-  ];
-  const sender = [
-    0, 0x20, 0, 0x20, 1, 0x20, 2, 0x10, 2, 0x0b,
-  ];
-  const code = section(10, [
-    8,
-    ...uleb(tick.length), ...tick,
-    ...uleb(cursor.length), ...cursor,
-    ...uleb(ui.length), ...ui,
-    ...uleb(command.length), ...command,
-    ...uleb(pair.length), ...pair,
-    ...uleb(frame.length), ...frame,
-    ...uleb(profession.length), ...profession,
-    ...uleb(sender.length), ...sender,
-  ]);
-  return Uint8Array.from([
-    0, 97, 115, 109, 1, 0, 0, 0,
-    ...type, ...imports, ...functions, ...table, ...memory, ...globals, ...exports,
-    ...elements, ...code,
-  ]);
-}
-
-/** A fixture function body, as the transform will read it. */
-function commandBody(bytes: Uint8Array, index: number): Uint8Array {
-  return parseCode(sectionById(splitSections(bytes), 10))[index]!;
-}
-
-function manifest(bytes: Uint8Array): KnownEnhancementBuild {
-  return {
-    sha256: createHash("sha256").update(bytes).digest("hex"),
-    outputSha256: PLACEHOLDER_OUTPUTS,
-    programId: 1,
-    buildId: 1,
-    hookFunction: 3,
-    hookParams: ["i32"],
-    hookResults: [],
-    tableSlot: 5,
-    commands: {
-      thunkExport: "enhancement_command",
-      professionTrace: {
-        readerExport: "enhancement_profession_trace",
-        sender: {
-          functionIndex: 10,
-          params: ["i32", "i32", "i32"],
-          results: [],
-          bodySha256: createHash("sha256")
-            .update(commandBody(bytes, 7))
-            .digest("hex"),
-        },
-      },
-      drain: {
-        functionIndex: 8,
-        params: ["i32", "i32"],
-        results: [],
-        tableSlot: 4,
-        bodySha256: createHash("sha256")
-          .update(commandBody(bytes, 5))
-          .digest("hex"),
-      },
-      entries: [{
-        opcode: 31,
-        functionIndex: 6,
-        params: ["i32"],
-        results: [],
-        bodySha256: createHash("sha256")
-          .update(commandBody(bytes, 3))
-          .digest("hex"),
-        label: "fixture command",
-      }, {
-        opcode: 93,
-        functionIndex: 7,
-        params: ["i32", "i32", "i32"],
-        results: [],
-        bodySha256: createHash("sha256")
-          .update(commandBody(bytes, 4))
-          .digest("hex"),
-        label: "fixture pair command",
-      }, {
-        opcode: 65,
-        functionIndex: 9,
-        params: ["i32", "i32"],
-        results: [],
-        bodySha256: createHash("sha256")
-          .update(commandBody(bytes, 6))
-          .digest("hex"),
-        label: "fixture profession command",
-      }],
-    },
-    cursorEvent: {
-      functionIndex: 4,
-      params: ["i32", "i32", "i32", "i32", "i32"],
-      results: [],
-      tableSlot: 1,
-      producerFunctions: [4, 4],
-    },
-    uiDispatcher: {
-      functionIndex: 5,
-      params: ["i32", "i32", "i32"],
-      results: [],
-      playerChatMessage: 0x1000_0082,
-      hideHeroPanelMessage: 0x1000_01a3,
-      showHeroPanelMessage: 0x1000_01a4,
-      partyDirtyMessages: PARTY_DIRTY_MESSAGES,
-      playerChatProducer: 5,
-      playerChatSites: 3,
-      nearbyPlayerMessages: [0x1000_007f, 0x1000_0080],
-      nearbyPlayerMessageProducers: [5, 5],
-    },
-    layout: {
-      contextRoot: 1, agentArray: 2, manualTargetAgentId: 3,
-      automaticTargetAgentId: 4, gameContextSlot: 6, characterContext: 4,
-      mapId: 5, isExplorable: 6, currentMapId: 7, currentInstanceType: 8,
-      playerNumber: 9, agentId: 10, agentX: 11, agentY: 12, agentType: 13,
-      agentPlayerNumber: 14, agentModelType: 15,
-      cursorActiveArt: 16, cursorSoftwareModel: 17, cursorShowCount: 18,
-      cursorColorBuffer: 19, cursorArtHotspot: 0, cursorArtTexture: 12,
-      cursorHandleKey: 8, cursorHandleObject: 0, cursorViewTexture: 8,
-      cursorTextureType: 12, cursorTextureWidth: 20, cursorTextureHeight: 24,
-      partyContext: 28, playerParty: 32, partyHeroes: 36,
-      heroMemberStride: 24, heroAgentId: 0, heroOwnerPlayerId: 4, heroId: 8,
-      // Distinct values rather than zeros: the config ABI is positional, so a
-      // fixture padded with zeros would let a mis-ordered field pass.
-      heroLevel: 42,
-      partyPlayers: 43, partyHenchmen: 44, partyFlag: 45,
-      accountContext: 78, accountUnlockedSkills: 79,
-      worldContext: 46, worldHeroFlags: 47, heroFlagStride: 48,
-      flagHeroId: 49, flagAgentId: 50, flagBehavior: 51,
-      worldHeroInfo: 52, heroInfoStride: 53, infoHeroId: 54,
-      infoAgentId: 40, infoLevel: 41,
-      infoPrimary: 55, infoSecondary: 56, infoAppearanceBitmap: 57,
-      worldSkillbars: 58, skillbarStride: 59, skillbarAgentId: 60,
-      skillbarSkills: 61, skillSlotStride: 62, skillSlotId: 63,
-      skillbarDisabled: 64,
-      worldAttributes: 65,
-      attributeStride: 66,
-      attributeAgentId: 67,
-      attributeEntries: 68,
-      attributeEntryStride: 69,
-      attributeEntryId: 70,
-      attributeEntryRank: 71,
-      areaInfo: 72,
-      areaInfoCount: 73,
-      areaInfoStride: 74,
-      areaInfoFlags: 75,
-      worldProfessionStates: 76,
-      professionStateStride: 77,
-      worldCharacterSkills: 80,
-    },
-  };
-}
-
-function callbackFixture(): Uint8Array {
-  const type = section(1, [
-    1, 0x60, 6, 0x7f, 0x7f, 0x7f, 0x7f, 0x7f, 0x7f, 0,
-  ]);
-  const imports = section(2, [
-    1, 3, 101, 110, 118, 8, 100, 105, 115, 112, 97, 116, 99, 104, 0, 0,
-  ]);
-  const functions = section(3, [1, 0]);
-  const name = [...new TextEncoder().encode("callback")];
-  const exports = section(7, [1, ...uleb(name.length), ...name, 0, 1]);
-  const body = [
-    0,
-    0x20, 0, 0x20, 1, 0x20, 2, 0x20, 3, 0x20, 4, 0x20, 5,
-    0x10, 0,
-    0x0b,
-  ];
-  const code = section(10, [1, ...uleb(body.length), ...body]);
-  return Uint8Array.from([
-    0, 97, 115, 109, 1, 0, 0, 0,
-    ...type, ...imports, ...functions, ...exports, ...code,
-  ]);
-}
-
+  callbackFixture,
+  CURSOR_ONLY,
+  CURSOR_TARGET,
+  CURSOR_TOOLBOX,
+  fixture,
+  manifest,
+  moduleWithManifest,
+  NO_CAPABILITIES,
+  PARTY_DIRTY_MESSAGES,
+  TARGET_ONLY,
+  UNSUPPORTED_ALL_CAPABILITIES,
+} from "../fixtures/enhancement-transform.js";
 describe("targeted Enhancement WebAssembly transform", () => {
   it("is deterministic, valid, and exports only the hook contract", () => {
     const input = fixture();
@@ -879,350 +534,6 @@ describe("targeted Enhancement WebAssembly transform", () => {
     ) as { configWords: number[] };
     evidence.configWords[1] = 1;
     assert.equal(decodeEnhancementManifest(moduleWithManifest(evidence)), null);
-  });
-
-  // The command queue is the entire write surface. These are the tests that
-  // decide whether this app can send a packet, so they instantiate the
-  // transformed module and drive the function rather than inspecting bytes.
-  it("emits the command queue for exactly the two certified command profiles", () => {
-    const input = fixture();
-    const build = manifest(input);
-    for (const capabilities of [CURSOR_ONLY, TARGET_ONLY, CURSOR_TARGET, CURSOR_TOOLBOX]) {
-      const output = transformEnhancementWasm(input, build, capabilities);
-      const exports = parseExports(sectionById(splitSections(output), 7));
-      assert.equal(
-        exports.some((entry) => entry.name === build.commands.thunkExport),
-        false,
-        "a read profile must carry no way to reach a packet builder at all",
-      );
-    }
-    for (const capabilities of [CURSOR_TOOLBOX_COMMANDS, CURSOR_TARGET_TOOLBOX_COMMANDS]) {
-      const output = transformEnhancementWasm(input, build, capabilities);
-      const exports = parseExports(sectionById(splitSections(output), 7));
-      assert.equal(
-        exports.filter((entry) => entry.name === build.commands.thunkExport).length,
-        1,
-      );
-      const module = new WebAssembly.Module(new Uint8Array(output));
-      assert.ok(decodeEnhancementManifest(module, capabilities));
-      assert.equal(
-        decodeEnhancementManifest(module, {
-          ...capabilities,
-          commands: false,
-        }),
-        null,
-        "a read-only profile must not accept a command-capable manifest",
-      );
-    }
-  });
-
-  it("dispatches queued commands only from the certified game-thread callback", () => {
-    const input = fixture();
-    const build = manifest(input);
-    const output = transformEnhancementWasm(input, build, CURSOR_TOOLBOX_COMMANDS);
-
-    const sent: number[] = [];
-    const instance = new WebAssembly.Instance(new WebAssembly.Module(new Uint8Array(output)), {
-      env: {
-        // The fixture's command function calls this; on the real client it is
-        // the packet builder.
-        t: (value: number) => { sent.push(value); },
-        c: () => {},
-        u: () => {},
-        tbl: new WebAssembly.Table({ initial: 6, maximum: 6, element: "anyfunc" }),
-      },
-    });
-    const command = instance.exports[build.commands.thunkExport] as
-      (opcode: number, a0: number, a1: number, a2: number, a3: number) => number;
-    const tick = instance.exports.EmscriptenExeThreadMainLoop as (value: number) => void;
-    const frame = instance.exports.frame as (value: number, context: number) => void;
-    const cursor = instance.exports.cursor as
-      (a: number, b: number, c: number, d: number, e: number) => void;
-    assert.equal(typeof command, "function");
-
-    assert.equal(command(31, 4242, 0, 0, 0), 1, "certified opcode is queued");
-    assert.deepEqual(sent, [], "enqueue never calls client code re-entrantly");
-    assert.equal(command(93, 11, 22, 33, 44), 0, "a pending command owns the mailbox");
-    cursor(1, 2, 3, 4, 5);
-    assert.deepEqual(sent, [], "an observer hook cannot dispatch commands");
-    tick(7);
-    assert.deepEqual(sent, [7], "the observer tick cannot dispatch commands");
-    frame(70, 700);
-    assert.deepEqual(
-      sent,
-      [7, 4242, 70],
-      "the command runs before the original frame boundary",
-    );
-
-    // A command taking more than one argument gets all of them, in order, and
-    // the arguments past its arity are ignored rather than passed on.
-    assert.equal(command(93, 11, 22, 33, 44), 1);
-    assert.deepEqual(sent, [7, 4242, 70], "the second command also waits");
-    tick(8);
-    assert.deepEqual(sent, [7, 4242, 70, 8]);
-    frame(80, 800);
-    assert.deepEqual(sent, [7, 4242, 70, 8, 80]);
-
-    assert.equal(command(31, 99, 0, 0, 0), 1);
-    assert.equal(command(0, 0, 0, 0, 0), 1, "opcode zero cancels the pending command");
-    frame(90, 900);
-    assert.deepEqual(sent, [7, 4242, 70, 8, 80, 90]);
-
-    // Everything else, including the neighbours of the ones we certified: the
-    // opcodes are a dense range on the real client, so an off-by-one would
-    // otherwise land on a real message.
-    sent.length = 0;
-    assert.equal(command(0, 4242, 0, 0, 0), 1, "opcode zero also cancels an empty mailbox");
-    for (const opcode of [1, 16, 21, 30, 32, 92, 94, -1, 0x7fff_ffff]) {
-      assert.equal(command(opcode, 4242, 0, 0, 0), 0, `opcode ${opcode}`);
-    }
-    assert.deepEqual(sent, [], "and nothing was sent");
-  });
-
-  it("distinguishes native and GWonMac command packets without changing them", () => {
-    const input = fixture();
-    const build = manifest(input);
-    const output = transformEnhancementWasm(input, build, CURSOR_TOOLBOX_COMMANDS);
-    const packets: number[][] = [];
-    const instance = new WebAssembly.Instance(
-      new WebAssembly.Module(new Uint8Array(output)),
-      {
-        env: {
-          t: () => {},
-          c: () => {},
-          u: (connection: number, size: number, pointer: number) => {
-            packets.push([connection, size, pointer]);
-          },
-        },
-      },
-    );
-    const wasmMemory = instance.exports.memory as WebAssembly.Memory;
-    const nativeProfession = instance.exports.profession as
-      (target: number, profession: number) => void;
-    const nativeSkill = instance.exports.skill as
-      (target: number, count: number, skills: number) => void;
-    const nativeSender = instance.exports.sender as
-      (connection: number, size: number, pointer: number) => void;
-    const enqueue = instance.exports[build.commands.thunkExport] as
-      (opcode: number, a0: number, a1: number, a2: number, a3: number) => number;
-    const frame = instance.exports.frame as (value: number, context: number) => void;
-    const readTrace = instance.exports[build.commands.professionTrace.readerExport] as
-      (pointer: number) => number;
-    const trace = () => {
-      assert.equal(readTrace(64), 30);
-      return [...new Uint32Array(wasmMemory.buffer, 64, 30)];
-    };
-
-    new Uint32Array(wasmMemory.buffer, 0, 2).set([31, 38]);
-    nativeSender(999, 8, 0);
-    assert.deepEqual(trace(), [
-      1,
-      0, 0, 0, 0,
-      0, 0, 0, 0,
-      1, 0, 999, 0, 0, 0, 0, 0, 0, 8,
-      31, 38, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    ]);
-
-    nativeProfession(77, 2);
-    assert.deepEqual(packets, [[999, 8, 0], [999, 12, 0]]);
-    assert.deepEqual([...new Uint32Array(wasmMemory.buffer, 0, 3)], [65, 77, 2]);
-    assert.deepEqual(trace(), [
-      1,
-      1, 0, 77, 2,
-      0, 0, 0, 0,
-      2, 0, 999, 0, 0, 0, 0, 0, 0, 12,
-      65, 77, 2, 0, 0, 0, 0, 0, 0, 0, 0,
-    ]);
-
-    assert.equal(enqueue(65, 88, 3, 0, 0), 1);
-    frame(1, 2);
-    assert.deepEqual(packets, [
-      [999, 8, 0],
-      [999, 12, 0],
-      [999, 12, 0],
-    ]);
-    assert.deepEqual([...new Uint32Array(wasmMemory.buffer, 0, 3)], [65, 88, 3]);
-    assert.deepEqual(trace(), [
-      1,
-      2, 1, 88, 3,
-      0, 0, 0, 0,
-      3, 1, 999, 0, 0, 0, 0, 0, 0, 12,
-      65, 88, 3, 0, 0, 0, 0, 0, 0, 0, 0,
-    ]);
-
-    const skillWords = new Uint32Array(wasmMemory.buffer, 256, 8);
-    skillWords.set([1, 2, 3, 4, 5, 6, 446, 8]);
-    nativeSkill(77, 8, 256);
-    assert.deepEqual(trace(), [
-      1,
-      2, 1, 88, 3,
-      1, 0, 77, 8,
-      4, 0, 999, 0, 0, 0, 0, 0, 0, 44,
-      93, 77, 8, 1, 2, 3, 4, 5, 6, 446, 8,
-    ]);
-    assert.equal(enqueue(93, 77, 8, 256, 0), 1);
-    frame(3, 4);
-    assert.deepEqual([...new Uint32Array(wasmMemory.buffer, 128, 11)], [
-      93, 77, 8, 1, 2, 3, 4, 5, 6, 446, 8,
-    ]);
-    assert.deepEqual(trace(), [
-      1,
-      2, 1, 88, 3,
-      2, 1, 77, 8,
-      5, 1, 999, 0, 0, 0, 0, 0, 0, 44,
-      93, 77, 8, 1, 2, 3, 4, 5, 6, 446, 8,
-    ]);
-  });
-
-  it("refuses a command whose function is not the one certified", () => {
-    const input = fixture();
-    const build = manifest(input);
-    // The failure this whole surface was reshaped around: an index that has
-    // drifted onto some other function. The type still matches -- the fixture's
-    // tick hook is also `(i32) -> void` -- so only the body hash catches it.
-    assert.throws(
-      () => transformEnhancementWasm(
-        input,
-        {
-          ...build,
-          commands: {
-            ...build.commands,
-            entries: [{ ...build.commands.entries[0]!, functionIndex: 3 }],
-          },
-        },
-        CURSOR_TOOLBOX_COMMANDS,
-      ),
-      /whose body is .* and not the certified/,
-    );
-    assert.throws(
-      () => transformEnhancementWasm(
-        input,
-        {
-          ...build,
-          commands: {
-            ...build.commands,
-            entries: [{ ...build.commands.entries[0]!, functionIndex: 4 }],
-          },
-        },
-        CURSOR_TOOLBOX_COMMANDS,
-      ),
-      /command opcode 31 signature is/,
-    );
-    assert.throws(
-      () => transformEnhancementWasm(
-        input,
-        {
-          ...build,
-          commands: {
-            ...build.commands,
-            entries: [{ ...build.commands.entries[0]!, functionIndex: 9_999 }],
-          },
-        },
-        CURSOR_TOOLBOX_COMMANDS,
-      ),
-      /out of range/,
-    );
-  });
-
-  it("refuses an uncertified or shared command drain boundary", () => {
-    const input = fixture();
-    const build = manifest(input);
-    assert.throws(
-      () => transformEnhancementWasm(
-        input,
-        {
-          ...build,
-          commands: {
-            ...build.commands,
-            drain: { ...build.commands.drain, bodySha256: "0".repeat(64) },
-          },
-        },
-        CURSOR_TOOLBOX_COMMANDS,
-      ),
-      /command drain boundary resolves .* not the certified/,
-    );
-    assert.throws(
-      () => transformEnhancementWasm(
-        input,
-        {
-          ...build,
-          commands: {
-            ...build.commands,
-            // The profession command has the same signature. Identity must still be
-            // distinct so command execution and scheduling cannot merge.
-            drain: {
-              ...build.commands.drain,
-              functionIndex: 9,
-              bodySha256: build.commands.entries[2]!.bodySha256,
-            },
-          },
-        },
-        CURSOR_TOOLBOX_COMMANDS,
-      ),
-      /must be distinct from hooks, commands, and sender/,
-    );
-    assert.throws(
-      () => transformEnhancementWasm(
-        input,
-        {
-          ...build,
-          commands: {
-            ...build.commands,
-            drain: { ...build.commands.drain, tableSlot: 0 },
-          },
-        },
-        CURSOR_TOOLBOX_COMMANDS,
-      ),
-      /command drain table slot 0 does not map/,
-    );
-  });
-
-  it("refuses an uncertified or shared traced packet sender", () => {
-    const input = fixture();
-    const build = manifest(input);
-    assert.throws(
-      () => transformEnhancementWasm(
-        input,
-        {
-          ...build,
-          commands: {
-            ...build.commands,
-            professionTrace: {
-              ...build.commands.professionTrace,
-              sender: {
-                ...build.commands.professionTrace.sender,
-                bodySha256: "0".repeat(64),
-              },
-            },
-          },
-        },
-        CURSOR_TOOLBOX_COMMANDS,
-      ),
-      /traced packet sender resolves .* not the certified/,
-    );
-    assert.throws(
-      () => transformEnhancementWasm(
-        input,
-        {
-          ...build,
-          commands: {
-            ...build.commands,
-            professionTrace: {
-              ...build.commands.professionTrace,
-              sender: {
-                ...build.commands.professionTrace.sender,
-                functionIndex: 5,
-                bodySha256: createHash("sha256")
-                  .update(commandBody(input, 2))
-                  .digest("hex"),
-              },
-            },
-          },
-        },
-        CURSOR_TOOLBOX_COMMANDS,
-      ),
-      /traced packet sender must be distinct from hooks and commands/,
-    );
   });
 
   it("checks the existing cursor table relation only when cursor is selected", () => {
